backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
supabase
python-multipart
numpy
python-dotenv
pydantic-settings
insightface
onnxruntime
faiss-cpu
scikit-image
